[qmimodem] Fix dialing to an international number
We can't just use ph->number because it has no leading '+' which is
crucial for a call (the operator says "The called number is not valid"
in case of a number without the plus).

Use phone_number_to_string() to get the valid number.

From: https://gitlab.manjaro.org/manjaro-arm/packages/community/plamo-gear/ofono/-/blob/master/0005-qmimodem-Fix-dialing-to-an-international-number.patch