udevng: Detect Quectel devices that use qmi_wwan_q driver
The qmi_wwan_q driver sets up pre-multiplexed network devices that
are children of the main device.  There can be up to 4 such devices,
but the exact number depends on the kernel configuration.  Detect and
report such devices to the modem driver by setting
"NumPremuxInterfaces" and "PremuxInterface{1..n}" properties.

For pre-multiplexed devices, we need to correlate the created order
of the device to the mux-id required by WDS_BIND_MUX_DATA_PORT_REQ.
For qmi_wwan_q devices the mux-id starts at 0x81.  So 0x81 for first
interface created, 0x82 for second, etc. For now we assume the mux-id
from the last digit of the interface name and set the
"PremuxInterface{1..n}MuxId" accordingly.

Also record the aggregation size for ul/dl from the driver qmap_size
parameter.