udevng: Unify set up of qmi based gobi driver
QMI modems need some additional information in order to be setup properly,
particular for newer devices.  The basic QMI setup consists of a 'cdc-wdm'
qmi device and a network interface.  The current detection logic works fine
for such devices.

When multiplexing is used, additional attributes need to be obtained in
order to send certain QMI WDA/WDS commands to bind the multiplexed
interfaces to a WDS session properly.

Rework the detection logic to use common QMI setup logic and record
additional attributes that will be needed later for QMI + multiplexing
setups.  The following attributes are tracked:
	"Bus" -> "usb", "pcie"
	"KernelDriver" -> driver providing the device (e.g. "qmi_wwan")
	"InterfaceNumber" -> interface number of the network interface